Update gem with Product classes for ProductTypes
Also changes UUIDs for S3 and RegisteredProvider to avoid conflicting
with UUIDs from JellyfishDemo extension.

Addresses https://github.com/boozallen/projectjellyfish/issues/1133.
