+++ app/models/jellyfish_aws/product/s3.rb
@@ -1,0 +1,14 @@
+module JellyfishAws
+  module Product
+    class S3 < ::Product
+      def order_questions
+        [
+        ]
+      end
+
+      def service_class
+        'JellyfishAws::Service::S3'.constantize
+      end
+    end
+  end
+end
